use two-stages param parser
