+++ pkg/server/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "conn.go",
         "conn_stmt.go",
+        "conn_stmt_params.go",
         "driver.go",
         "driver_tidb.go",
         "extension.go",
@@ -32,6 +33,7 @@ go_library(
         "//pkg/infoschema",
         "//pkg/kv",
         "//pkg/metrics",
+        "//pkg/param",
         "//pkg/parser",
         "//pkg/parser/ast",
         "//pkg/parser/auth",
@@ -76,6 +78,7 @@ go_library(
         "//pkg/util/arena",
         "//pkg/util/chunk",
         "//pkg/util/cpuprofile",
+        "//pkg/util/dbterror",
         "//pkg/util/dbterror/exeerrors",
         "//pkg/util/execdetails",
         "//pkg/util/fastrand",
@@ -125,6 +128,7 @@ go_test(
     name = "server_test",
     timeout = "short",
     srcs = [
+        "conn_stmt_params_test.go",
         "conn_stmt_test.go",
         "conn_test.go",
         "driver_tidb_test.go",
@@ -138,20 +142,23 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":server"],
     flaky = True,
-    shard_count = 48,
+    shard_count = 50,
     deps = [
         "//pkg/config",
         "//pkg/domain",
         "//pkg/domain/infosync",
+        "//pkg/expression",
         "//pkg/extension",
         "//pkg/keyspace",
         "//pkg/kv",
         "//pkg/metrics",
+        "//pkg/param",
         "//pkg/parser/ast",
         "//pkg/parser/auth",
         "//pkg/parser/charset",
         "//pkg/parser/model",
         "//pkg/parser/mysql",
+        "//pkg/parser/terror",
         "//pkg/server/internal",
         "//pkg/server/internal/column",
         "//pkg/server/internal/handshake",
