+++ pkg/server/extension.go
@@ -17,8 +17,8 @@ package server
 import (
 	"fmt"
 
-	"github.com/pingcap/tidb/pkg/expression"
 	"github.com/pingcap/tidb/pkg/extension"
+	"github.com/pingcap/tidb/pkg/param"
 	"github.com/pingcap/tidb/pkg/parser"
 	"github.com/pingcap/tidb/pkg/parser/ast"
 	"github.com/pingcap/tidb/pkg/parser/auth"
@@ -57,7 +57,7 @@ func (cc *clientConn) onExtensionConnEvent(tp extension.ConnEventTp, err error) 
 	cc.extensions.OnConnectionEvent(tp, info)
 }
 
-func (cc *clientConn) onExtensionStmtEnd(node interface{}, stmtCtxValid bool, err error, args ...expression.Expression) {
+func (cc *clientConn) onExtensionStmtEnd(node interface{}, stmtCtxValid bool, err error, args ...param.BinaryParam) {
 	if !cc.extensions.HasStmtEventListeners() {
 		return
 	}
@@ -85,9 +85,17 @@ func (cc *clientConn) onExtensionStmtEnd(node interface{}, stmtCtxValid bool, er
 	case PreparedStatement:
 		info.executeStmtID = uint32(stmt.ID())
 		prepared, _ := sessVars.GetPreparedStmtByID(info.executeStmtID)
+
+		// TODO: the `BinaryParam` is parsed two times: one in the `Execute` method and one here. It would be better to
+		// eliminate one of them by storing the parsed result.
+		typectx := ctx.GetSessionVars().StmtCtx.TypeCtx()
+		typectx = types.NewContext(typectx.Flags(), typectx.Location(), func(_ error) {
+			// ignore all warnings
+		})
+		params, _ := param.ExecArgs(typectx, args)
 		info.executeStmt = &ast.ExecuteStmt{
 			PrepStmt:   prepared,
-			BinaryArgs: args,
+			BinaryArgs: params,
 		}
 		info.stmtNode = info.executeStmt
 	case ast.StmtNode:
@@ -115,7 +123,7 @@ func (cc *clientConn) onExtensionSQLParseFailed(sql string, err error) {
 	})
 }
 
-func (cc *clientConn) onExtensionBinaryExecuteEnd(prep PreparedStatement, args []expression.Expression, stmtCtxValid bool, err error) {
+func (cc *clientConn) onExtensionBinaryExecuteEnd(prep PreparedStatement, args []param.BinaryParam, stmtCtxValid bool, err error) {
 	cc.onExtensionStmtEnd(prep, stmtCtxValid, err, args...)
 }
