+++ pkg/server/internal/parse/BUILD.bazel
@@ -6,15 +6,9 @@ go_library(
     importpath = "github.com/pingcap/tidb/pkg/server/internal/parse",
     visibility = ["//pkg/server:__subpackages__"],
     deps = [
-        "//pkg/errno",
-        "//pkg/expression",
-        "//pkg/parser/charset",
         "//pkg/parser/mysql",
         "//pkg/server/internal/handshake",
         "//pkg/server/internal/util",
-        "//pkg/types",
-        "//pkg/util/dbterror",
-        "//pkg/util/hack",
         "//pkg/util/logutil",
         "@com_github_klauspost_compress//zstd",
         "@org_uber_go_zap//:zap",
@@ -30,14 +24,9 @@ go_test(
     ],
     embed = [":parse"],
     flaky = True,
-    shard_count = 4,
     deps = [
-        "//pkg/expression",
         "//pkg/parser/mysql",
-        "//pkg/parser/terror",
         "//pkg/server/internal/handshake",
-        "//pkg/server/internal/util",
-        "//pkg/types",
         "@com_github_stretchr_testify//require",
     ],
 )
