+++ pkg/server/internal/parse/parse.go
@@ -18,380 +18,19 @@ import (
 	"bytes"
 	"context"
 	"encoding/binary"
-	"fmt"
-	"math"
 
 	"github.com/klauspost/compress/zstd"
-	"github.com/pingcap/tidb/pkg/errno"
-	"github.com/pingcap/tidb/pkg/expression"
-	"github.com/pingcap/tidb/pkg/parser/charset"
 	"github.com/pingcap/tidb/pkg/parser/mysql"
 	"github.com/pingcap/tidb/pkg/server/internal/handshake"
 	util2 "github.com/pingcap/tidb/pkg/server/internal/util"
-	"github.com/pingcap/tidb/pkg/types"
-	"github.com/pingcap/tidb/pkg/util/dbterror"
-	"github.com/pingcap/tidb/pkg/util/hack"
 	"github.com/pingcap/tidb/pkg/util/logutil"
 	"go.uber.org/zap"
 )
-
-var errUnknownFieldType = dbterror.ClassServer.NewStd(errno.ErrUnknownFieldType)
 
 // maxFetchSize constants
 const (
 	maxFetchSize = 1024
 )
-
-// ExecArgs parse execute arguments to datum slice.
-func ExecArgs(typectx types.Context, params []expression.Expression, boundParams [][]byte,
-	nullBitmap, paramTypes, paramValues []byte, enc *util2.InputDecoder) (err error) {
-	pos := 0
-	var (
-		tmp    interface{}
-		v      []byte
-		n      int
-		isNull bool
-	)
-	if enc == nil {
-		enc = util2.NewInputDecoder(charset.CharsetUTF8)
-	}
-
-	args := make([]types.Datum, len(params))
-	for i := 0; i < len(args); i++ {
-		// if params had received via ComStmtSendLongData, use them directly.
-		// ref https://dev.mysql.com/doc/internals/en/com-stmt-send-long-data.html
-		// see clientConn#handleStmtSendLongData
-		if boundParams[i] != nil {
-			args[i] = types.NewBytesDatum(enc.DecodeInput(boundParams[i]))
-			continue
-		}
-
-		// check nullBitMap to determine the NULL arguments.
-		// ref https://dev.mysql.com/doc/internals/en/com-stmt-execute.html
-		// notice: some client(e.g. mariadb) will set nullBitMap even if data had be sent via ComStmtSendLongData,
-		// so this check need place after boundParam's check.
-		if nullBitmap[i>>3]&(1<<(uint(i)%8)) > 0 {
-			var nilDatum types.Datum
-			nilDatum.SetNull()
-			args[i] = nilDatum
-			continue
-		}
-
-		if (i<<1)+1 >= len(paramTypes) {
-			return mysql.ErrMalformPacket
-		}
-
-		tp := paramTypes[i<<1]
-		isUnsigned := (paramTypes[(i<<1)+1] & 0x80) > 0
-
-		switch tp {
-		case mysql.TypeNull:
-			var nilDatum types.Datum
-			nilDatum.SetNull()
-			args[i] = nilDatum
-			continue
-
-		case mysql.TypeTiny:
-			if len(paramValues) < (pos + 1) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-
-			if isUnsigned {
-				args[i] = types.NewUintDatum(uint64(paramValues[pos]))
-			} else {
-				args[i] = types.NewIntDatum(int64(int8(paramValues[pos])))
-			}
-
-			pos++
-			continue
-
-		case mysql.TypeShort, mysql.TypeYear:
-			if len(paramValues) < (pos + 2) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-			valU16 := binary.LittleEndian.Uint16(paramValues[pos : pos+2])
-			if isUnsigned {
-				args[i] = types.NewUintDatum(uint64(valU16))
-			} else {
-				args[i] = types.NewIntDatum(int64(int16(valU16)))
-			}
-			pos += 2
-			continue
-
-		case mysql.TypeInt24, mysql.TypeLong:
-			if len(paramValues) < (pos + 4) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-			valU32 := binary.LittleEndian.Uint32(paramValues[pos : pos+4])
-			if isUnsigned {
-				args[i] = types.NewUintDatum(uint64(valU32))
-			} else {
-				args[i] = types.NewIntDatum(int64(int32(valU32)))
-			}
-			pos += 4
-			continue
-
-		case mysql.TypeLonglong:
-			if len(paramValues) < (pos + 8) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-			valU64 := binary.LittleEndian.Uint64(paramValues[pos : pos+8])
-			if isUnsigned {
-				args[i] = types.NewUintDatum(valU64)
-			} else {
-				args[i] = types.NewIntDatum(int64(valU64))
-			}
-			pos += 8
-			continue
-
-		case mysql.TypeFloat:
-			if len(paramValues) < (pos + 4) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-
-			args[i] = types.NewFloat32Datum(math.Float32frombits(binary.LittleEndian.Uint32(paramValues[pos : pos+4])))
-			pos += 4
-			continue
-
-		case mysql.TypeDouble:
-			if len(paramValues) < (pos + 8) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-
-			args[i] = types.NewFloat64Datum(math.Float64frombits(binary.LittleEndian.Uint64(paramValues[pos : pos+8])))
-			pos += 8
-			continue
-
-		case mysql.TypeDate, mysql.TypeTimestamp, mysql.TypeDatetime:
-			if len(paramValues) < (pos + 1) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-			// See https://dev.mysql.com/doc/internals/en/binary-protocol-value.html
-			// for more details.
-			length := paramValues[pos]
-			pos++
-			switch length {
-			case 0:
-				tmp = types.ZeroDatetimeStr
-			case 4:
-				pos, tmp = binaryDate(pos, paramValues)
-			case 7:
-				pos, tmp = binaryDateTime(pos, paramValues)
-			case 11:
-				pos, tmp = binaryTimestamp(pos, paramValues)
-			case 13:
-				pos, tmp = binaryTimestampWithTZ(pos, paramValues)
-			default:
-				err = mysql.ErrMalformPacket
-				return
-			}
-			// TODO: generate the time datum directly
-			var parseTime func(types.Context, string) (types.Time, error)
-			switch tp {
-			case mysql.TypeDate:
-				parseTime = types.ParseDate
-			case mysql.TypeDatetime:
-				parseTime = types.ParseDatetime
-			case mysql.TypeTimestamp:
-				// To be compatible with MySQL, even the type of parameter is
-				// TypeTimestamp, the return type should also be `Datetime`.
-				parseTime = types.ParseDatetime
-			}
-			var time types.Time
-			time, err = parseTime(typectx, tmp.(string))
-			err = typectx.HandleTruncate(err)
-			if err != nil {
-				// TODO: handle the warning and errors
-				return
-			}
-			args[i] = types.NewDatum(time)
-			continue
-
-		case mysql.TypeDuration:
-			if len(paramValues) < (pos + 1) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-			// See https://dev.mysql.com/doc/internals/en/binary-protocol-value.html
-			// for more details.
-			length := paramValues[pos]
-			pos++
-			switch length {
-			case 0:
-				tmp = "0"
-			case 8:
-				isNegative := paramValues[pos]
-				if isNegative > 1 {
-					err = mysql.ErrMalformPacket
-					return
-				}
-				pos++
-				pos, tmp = binaryDuration(pos, paramValues, isNegative)
-			case 12:
-				isNegative := paramValues[pos]
-				if isNegative > 1 {
-					err = mysql.ErrMalformPacket
-					return
-				}
-				pos++
-				pos, tmp = binaryDurationWithMS(pos, paramValues, isNegative)
-			default:
-				err = mysql.ErrMalformPacket
-				return
-			}
-			var dur types.Duration
-			dur, _, err = types.ParseDuration(typectx, tmp.(string), types.MaxFsp)
-			err = typectx.HandleTruncate(err)
-			if err != nil {
-				return
-			}
-			args[i] = types.NewDatum(dur)
-			continue
-		case mysql.TypeNewDecimal:
-			if len(paramValues) < (pos + 1) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-
-			v, isNull, n, err = util2.ParseLengthEncodedBytes(paramValues[pos:])
-			pos += n
-			if err != nil {
-				return
-			}
-
-			if isNull {
-				args[i] = types.NewDecimalDatum(nil)
-			} else {
-				var dec types.MyDecimal
-				err = typectx.HandleTruncate(dec.FromString(v))
-				if err != nil {
-					return err
-				}
-				args[i] = types.NewDecimalDatum(&dec)
-			}
-			continue
-		case mysql.TypeBlob, mysql.TypeTinyBlob, mysql.TypeMediumBlob, mysql.TypeLongBlob:
-			if len(paramValues) < (pos + 1) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-			v, isNull, n, err = util2.ParseLengthEncodedBytes(paramValues[pos:])
-			pos += n
-			if err != nil {
-				return
-			}
-
-			if isNull {
-				args[i] = types.NewBytesDatum(nil)
-			} else {
-				args[i] = types.NewBytesDatum(v)
-			}
-			continue
-		case mysql.TypeUnspecified, mysql.TypeVarchar, mysql.TypeVarString, mysql.TypeString,
-			mysql.TypeEnum, mysql.TypeSet, mysql.TypeGeometry, mysql.TypeBit:
-			if len(paramValues) < (pos + 1) {
-				err = mysql.ErrMalformPacket
-				return
-			}
-
-			v, isNull, n, err = util2.ParseLengthEncodedBytes(paramValues[pos:])
-			pos += n
-			if err != nil {
-				return
-			}
-
-			if !isNull {
-				v = enc.DecodeInput(v)
-				tmp = string(hack.String(v))
-			} else {
-				tmp = nil
-			}
-			args[i] = types.NewDatum(tmp)
-			continue
-		default:
-			err = errUnknownFieldType.GenWithStack("stmt unknown field type %d", tp)
-			return
-		}
-	}
-
-	for i := range params {
-		ft := new(types.FieldType)
-		types.InferParamTypeFromUnderlyingValue(args[i].GetValue(), ft)
-		params[i] = &expression.Constant{Value: args[i], RetType: ft}
-	}
-	return
-}
-
-func binaryDate(pos int, paramValues []byte) (int, string) {
-	year := binary.LittleEndian.Uint16(paramValues[pos : pos+2])
-	pos += 2
-	month := paramValues[pos]
-	pos++
-	day := paramValues[pos]
-	pos++
-	return pos, fmt.Sprintf("%04d-%02d-%02d", year, month, day)
-}
-
-func binaryDateTime(pos int, paramValues []byte) (int, string) {
-	pos, date := binaryDate(pos, paramValues)
-	hour := paramValues[pos]
-	pos++
-	minute := paramValues[pos]
-	pos++
-	second := paramValues[pos]
-	pos++
-	return pos, fmt.Sprintf("%s %02d:%02d:%02d", date, hour, minute, second)
-}
-
-func binaryTimestamp(pos int, paramValues []byte) (int, string) {
-	pos, dateTime := binaryDateTime(pos, paramValues)
-	microSecond := binary.LittleEndian.Uint32(paramValues[pos : pos+4])
-	pos += 4
-	return pos, fmt.Sprintf("%s.%06d", dateTime, microSecond)
-}
-
-func binaryTimestampWithTZ(pos int, paramValues []byte) (int, string) {
-	pos, timestamp := binaryTimestamp(pos, paramValues)
-	tzShiftInMin := int16(binary.LittleEndian.Uint16(paramValues[pos : pos+2]))
-	tzShiftHour := tzShiftInMin / 60
-	tzShiftAbsMin := tzShiftInMin % 60
-	if tzShiftAbsMin < 0 {
-		tzShiftAbsMin = -tzShiftAbsMin
-	}
-	pos += 2
-	return pos, fmt.Sprintf("%s%+02d:%02d", timestamp, tzShiftHour, tzShiftAbsMin)
-}
-
-func binaryDuration(pos int, paramValues []byte, isNegative uint8) (int, string) {
-	sign := ""
-	if isNegative == 1 {
-		sign = "-"
-	}
-	days := binary.LittleEndian.Uint32(paramValues[pos : pos+4])
-	pos += 4
-	hours := paramValues[pos]
-	pos++
-	minutes := paramValues[pos]
-	pos++
-	seconds := paramValues[pos]
-	pos++
-	return pos, fmt.Sprintf("%s%d %02d:%02d:%02d", sign, days, hours, minutes, seconds)
-}
-
-func binaryDurationWithMS(pos int, paramValues []byte,
-	isNegative uint8) (int, string) {
-	pos, dur := binaryDuration(pos, paramValues, isNegative)
-	microSecond := binary.LittleEndian.Uint32(paramValues[pos : pos+4])
-	pos += 4
-	return pos, fmt.Sprintf("%s.%06d", dur, microSecond)
-}
 
 // StmtFetchCmd parse COM_STMT_FETCH command
 func StmtFetchCmd(data []byte) (stmtID uint32, fetchSize uint32, err error) {
