+++ pkg/session/BUILD.bazel
@@ -35,6 +35,7 @@ go_library(
         "//pkg/meta",
         "//pkg/metrics",
         "//pkg/owner",
+        "//pkg/param",
         "//pkg/parser",
         "//pkg/parser/ast",
         "//pkg/parser/auth",
